requirements.txt
altair~=5.1.1
numpy~=1.26.0
pandas~=2.1.1
pydeck~=0.8.0
streamlit~=1.27.0
google-cloud-bigquery==3.11.4
google-auth 
google-auth-httplib2 
google-api-python-client
db-dtypes